Fixed bug preventing apprpriate route matching:
the route productRouter.get("/:id", getProduct) will catch all GET requests with any string after the initial "/", including the paths "/search" and "/filter". This is because the :id parameter acts as a wildcard, matching any value at that position in the URL. fixed this by defining more specific routes before the generic ones.